feat: change data type

+++ src/main/kotlin/models/Item.kt
@@ -3,4 +3,4 @@ package models
 /**
  * This class represents an item to be carried by the rockets.
  */
-class Item (var name: String, var weight: Int)
+class Item (var name: String, var weight: Float)
